try-chainer/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
chainer==1.16.0
cycler==0.10.0            # via matplotlib
filelock==2.0.6           # via chainer
matplotlib==1.5.3
mecab-python3==0.7
nose==1.3.7               # via chainer
numpy==1.11.2
pillow==3.4.1
protobuf==3.1.0.post1     # via chainer
pyparsing==2.1.10         # via matplotlib
python-dateutil==2.5.3    # via matplotlib
pytz==2016.7              # via matplotlib
scikit-learn==0.18
scipy==0.18.1
six==1.10.0               # via chainer, cycler, protobuf, python-dateutil

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via protobuf
